+++ src/client/src/ui/spotTile/priceButton/PriceButton.tsx
@@ -3,8 +3,8 @@ import * as classnames from 'classnames'
 
 import { Direction } from '../../../types'
 import './PriceButtonStyles.scss'
-import { CurrencyPair } from "../../../types/currencyPair";
-import { Rate } from "../../../types/rate";
+import { CurrencyPair } from '../../../types/currencyPair'
+import { Rate } from '../../../types/rate'
 
 interface PriceButtonProps {
   className: string
